requirements.txt
django~=4.1.5
django-phonenumber-field==7.0.2
Pillow==9.4.0
yookassa==2.3.3
environs==9.5.0
django_htmx==1.13.0
